requirements.txt
streamlit
numpy
scikit-learn
xgboost
gspread
oauth2client
google-auth
joblib